requirements.txt
gitdb==4.0.7
GitPython==3.1.17
protobuf==3.15.8
six==1.15.0
smmap==4.0.0
